+++ TabletopTweaks/NewContent/FighterAdvancedArmorTrainings/AdvancedArmorTraining.cs
@@ -2,6 +2,7 @@
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Prerequisites;
 using Kingmaker.Blueprints.Classes.Selection;
+using Kingmaker.Designers.Mechanics.Facts;
 using System;
 using TabletopTweaks.Config;
 using TabletopTweaks.Extensions;
@@ -15,6 +16,15 @@ namespace TabletopTweaks.NewContent.FighterAdvancedArmorTrainings {
             var ArmorTraining = Resources.GetBlueprint<BlueprintFeature>("3c380607706f209499d951b29d3c44f3");
             var FighterFeatSelection = Resources.GetBlueprint<BlueprintFeatureSelection>("41c8486641f7d6d4283ca9dae4147a9f");
 
+            var ArmorTrainingSpeedFeature = Helpers.CreateBlueprint<BlueprintFeature>("ArmorTrainingSpeedFeature", bp => {
+                bp.Ranks = 5;
+                bp.SetName("Armor Training Speed");
+                bp.SetDescription("");
+                bp.HideInUI = true;
+                bp.HideInCharacterSheetAndLevelUp = true;
+                bp.IsClassFeature = true;
+                bp.AddComponent<ArmorSpeedPenaltyRemoval>();
+            });
             var AdvancedArmorTrainingSelection = Helpers.CreateBlueprint<BlueprintFeatureSelection>("AdvancedArmorTrainingSelection", bp => {
                 bp.Ranks = 3;
                 bp.SetName("Advanced Armor Training");